Fixed head order
CSS files should be called before everything else in the head tag to prevent showing the contents of the page without the css having been applied. Rule of thumb: CSS FIRST.
Also removed empty script tag as this is a hack that will likely be fixed and may not work on every browser anyways.